requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.1
gunicorn==21.2.0
plotly==5.18.0
pandas==2.2.0 